refactor(workflow): extract workflow event payload builders to a module
The payload builders (statement events, readiness-blocker, report-state) are
pure WorkflowEventCreate constructors with no DB/session logic. Move them, plus
their only deps (build_workflow_dedupe_key, PACKAGE_WORKFLOW_SOURCE_ID), to
services/workflow_event_builders.py. workflow_events.py imports them back
(re-exported) so call sites are unchanged; no cycle (builders import only
enums/schemas/models).

workflow_events.py: 902 -> 764. Structural. 28 workflow tests green; ruff clean.
